pyproject.toml
[tool.poetry]
name = "discord-py-framework-1"
version = "0.1.0"
description = ""
authors = ["Multi76 <<>>"]

[tool.poetry.dependencies]
python = "^3.8"
discord = "^1.0.1"
PyYAML = "^5.4.1"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
